server/pyproject.toml
[project]
name = "shelter-volunteers-server"
version = "0.1.0"
description = "Server for Shelter Volunteers application"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "flask>=3.1.2",
    "flask-cors>=6.0.1",
    "google-auth>=2.43.0",
    "google-auth-oauthlib>=1.2.2",
    "gunicorn>=23.0.0",
    "pyjwt>=2.10.1",
    "pymongo>=4.15.4",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
]

[dependency-groups]
dev = [
    "pylint>=3.3.9",
    "pytest>=8.4.2",
    "testcontainers[mongodb]>=4.13.0",
]
